get stamp id
